Removed the "plugins_" introducer.
I changed the name of all the classes so they don't start with `plugins_...`.
It's alrady inside the "serverplugins" namespace so the repeat is not that
useful, although that means I had to change a few function and variable
names that were clashing with the new shorter names (i.e. there was a
`names()` function and a `names` variable, for example).