Added a server identifier.
Created an atom like identifier feature so we can have up to 255 servers with plugins.
Prevent addition of "server" as a dependency (that is, we know that a server is always available, so there is no need to have it as a dependency and the push() fails with that name).
Removed the X_OK test in the to_filename() function. .so files are not saved with the execute flag set.