Fixed the SERVERPLUGINS_LISTEN macros.
The new version of the server plugins supports a list of names, a collection,
a server, and plugins, all of which are well interconnected. The old
macros assumed that we needed an instance() function in a plugin to be
able to access it. This is not the case anymore. We can now use the
collection, which each plugin receives automatically, to find another
plugin and access it if loaded.