Tweaked the library to avoid "instance()".
The library used a strange map to save the servers. Looking at it closer,
I could see that was not required so I decided to remove it. The server
is already saved in the collection at the time we do the load_plugins()
which is more than sufficient since we can just do get_server().
And for plugins, there is also no need for an instance() since we can
call get_plugin() on the collection and the collection is accessible
from all the plugins using the plugins() function.

So as a result I removed all of that weirdness. The initialization is
just a tad bit _broken_ for the server plugin as it requires us to have
the complete_plugin_initialization() to save the plugin shared pointer
in its factory.

Also I think that it is already way cleaner than what I had in the old
days (quite different from the libsnapwebsites implementation).