Remove the emitter's class namespace.
When accessing a plugin from another one, the namespaces have to be specified
only the plugin name does not include the namespace (i.e. the plugin defined
as `communicatord::loadavg::loadavg` is named `loadavg`). Now the macros used
to listen to that plugin properly ignore all the namespaces.

Added a note with the concept of using templated functions instead of macros
but so far I have no clue how that would be done since we do use functionality
specific to macros (the '#' and '##' operators).

Added a LISTEN macro that allows us to specify the callback directly so
that way we are not forced to name it "on_<signal>" and can use something
other than an std::bind() such as a lambda if required.